+++ stxt/src/extension/CompletionProviderSearch.ts
@@ -46,7 +46,7 @@ export function buscarSugerenciasPorParent(parent: Node, prefix: string): Comple
         }
         item.detail = childName;
 
-        const actualChildren: Node[] = parent.getChildrenByName(childDef.getName());
+        const actualChildren: Node[] = parent.getChildrenByName(childDef.getName(), childDef.getNamespace());
         const maxChilds = childDef.getMax() ?? -1;
         if (maxChilds < 0 || actualChildren.length < maxChilds) {
             result.push(item);